requirements.txt
absl-py==0.15.0
aiohttp==3.8.1
aiosignal==1.2.0
alabaster==0.7.12
albumentations==1.1.0
anaconda-client==1.9.0
anaconda-navigator==2.1.1
anaconda-project==0.10.1
anyio==2.2.0
appdirs==1.4.4
argh==0.26.2
argon2-cffi==20.1.0
arrow==0.13.1
asn1crypto==1.4.0
astroid==2.6.6
astropy==4.3.1
async-generator==1.10
async-timeout==4.0.1
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.5.7
babel==2.9.1
backcall==0.2.0
backports.functools-lru-cache==1.6.4
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.2.0
beautifulsoup4==4.10.0
binaryornot==0.4.4
bitarray==2.3.0
bkcharts==0.2
black==19.10b0
bleach==4.0.0
blinker==1.4
bokeh==2.4.1
boto==2.49.0
bottleneck==1.3.2
brotlipy==0.7.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.10.8
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.3
cloudpickle==2.0.0
clyent==1.2.2
colorama==0.4.4
comtypes==1.1.10
conda-build==3.21.6
conda-content-trust==0+unknown
conda-pack==0.6.0
conda-package-handling==1.7.3
conda-repo-cli==1.0.4
conda-token==0.3.0
conda-verify==3.4.2
conda==4.11.0
contextlib2==0.6.0.post1
cookiecutter==1.7.2
cryptography==3.4.8
cycler==0.10.0
cython==0.29.24
cytoolz==0.11.0
daal4py==2021.3.0
dask==2021.10.0
debugpy==1.4.1
decorator==5.1.0
defusedxml==0.7.1
diff-match-patch==20200713
distributed==2021.10.0
docker-pycreds==0.4.0
docutils==0.17.1
entrypoints==0.3
et-xmlfile==1.1.0
fastcache==1.1.0
filelock==3.3.1
flake8==3.9.2
flask==1.1.2
fonttools==4.25.0
frozenlist==1.2.0
fsspec==2021.10.1
future==0.18.2
gevent==21.8.0
gitdb==4.0.9
gitpython==3.1.27
glob2==0.7
google-auth-oauthlib==0.4.1
google-auth==1.35.0
greenlet==1.1.1
grpcio==1.42.0
h5py==3.2.1
heapdict==1.0.1
html5lib==1.1
idna==3.2
imagecodecs==2021.8.26
imageio==2.9.0
imagesize==1.2.0
importlib-metadata==4.8.1
inflection==0.5.1
iniconfig==1.1.1
intervaltree==3.1.0
ipykernel==6.4.1
ipython-genutils==0.2.0
ipython==7.29.0
ipywidgets==7.6.5
isort==5.9.3
itsdangerous==2.0.1
jdcal==1.4.1
jedi==0.18.0
jinja2-time==0.2.0
jinja2==2.11.3
joblib==1.1.0
json5==0.9.6
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyter-server==1.4.1
jupyter==1.0.0
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
jupyterlab-widgets==1.0.0
jupyterlab==3.2.1
keyring==23.1.0
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
libarchive-c==2.9
llvmlite==0.37.0
locket==0.2.1
lxml==4.6.3
markdown==3.3.4
markupsafe==1.1.1
matplotlib-inline==0.1.2
matplotlib==3.4.3
mccabe==0.6.1
menuinst==1.4.18
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mock==4.0.3
more-itertools==8.10.0
mpmath==1.2.1
msgpack==1.0.2
multidict==5.1.0
multipledispatch==0.6.0
munkres==1.1.4
mypy-extensions==0.4.3
navigator-updater==0.2.1
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
nltk==3.6.5
nose==1.3.7
notebook==6.4.5
numba==0.54.1
numexpr==2.7.3
numpy==1.20.3
numpydoc==1.1.0
oauthlib==3.2.0
olefile==0.46
opencv-python-headless==4.5.5.64
opencv-python==4.5.5.64
openpyxl==3.0.9
packaging==21.0
pandas==1.3.4
pandocfilters==1.4.3
paramiko==2.7.2
parso==0.8.2
partd==1.2.0
path==16.0.0
pathlib2==2.3.6
pathspec==0.7.0
pathtools==0.1.2
patsy==0.5.2
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
pillow==8.4.0
pip==21.2.4
pkginfo==1.7.1
pluggy==0.13.1
ply==3.11
poyo==0.5.0
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.20
protobuf==3.19.1
psutil==5.8.0
ptyprocess==0.7.0
py==1.10.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycodestyle==2.7.0
pycosat==0.6.3
pycparser==2.20
pycurl==7.44.1
pydocstyle==6.1.1
pyerfa==2.0.0
pyflakes==2.3.1
pygments==2.10.0
pyjwt==2.1.0
pylint==2.9.6
pyls-spyder==0.4.0
pynacl==1.4.0
pyodbc==4.0.0-unsupported
pyopenssl==21.0.0
pyparsing==3.0.4
pyreadline==2.1
pyrsistent==0.18.0
pysocks==1.7.1
pytest==6.2.4
python-dateutil==2.8.2
python-lsp-black==1.0.0
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.2.4
python-slugify==5.0.2
pytz==2021.3
pywavelets==1.1.1
pywin32-ctypes==0.2.0
pywin32==228
pywinpty==0.5.7
pyyaml==6.0
pyzmq==22.2.1
qdarkstyle==3.0.2
qstylizer==0.1.10
qtawesome==1.0.2
qtconsole==5.1.1
qtpy==1.10.0
qudida==0.0.4
regex==2021.8.3
requests-oauthlib==1.3.0
requests==2.26.0
rope==0.19.0
rsa==4.7.2
rtree==0.9.7
ruamel-yaml-conda==0.15.100
scikit-image==0.18.3
scikit-learn-intelex==2021.20210714.120553
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.2
send2trash==1.8.0
sentry-sdk==1.5.7
setproctitle==1.2.2
setuptools==58.0.4
shortuuid==1.0.8
simplegeneric==0.8.1
singledispatch==3.7.0
sip==4.19.13
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
snowballstemmer==2.1.0
sortedcollections==2.1.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sphinx==4.2.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sphinxcontrib-websupport==1.2.4
spyder-kernels==2.1.3
spyder==5.1.5
sqlalchemy==1.4.22
statsmodels==0.12.2
sympy==1.9
tables==3.6.1
tbb==0.2
tblib==1.7.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.6.0
tensorboard==2.6.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.5.0
text-unidecode==1.3
textdistance==4.2.1
threadpoolctl==2.2.0
three-merge==0.1.1
tifffile==2021.7.2
tinycss==0.4
toml==0.10.2
toolz==0.11.1
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
typed-ast==1.4.3
typing-extensions==3.10.0.2
ujson==4.0.2
unicodecsv==0.14.1
unidecode==1.2.0
urllib3==1.26.7
wandb==0.12.11
watchdog==2.1.3
wcwidth==0.2.5
webencodings==0.5.1
werkzeug==2.0.2
wheel==0.37.0
whichcraft==0.6.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.12.1
xlrd==2.0.1
xlsxwriter==3.0.1
xlwings==0.24.9
xlwt==1.3.0
xmltodict==0.12.0
yapf==0.31.0
yarl==1.6.3
yaspin==2.1.0
zict==2.0.0
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0